Added REAME Files
Motivation
---
Provide README files that provide an overview of the sample files used in a typical CICD process. The goal is to help newcomers get up to speed quickly on the steps and process necessary to automate the build, deployment, and testing of new code in a CICD environment.

Motification
---
- Added a Project Readme
- Added a Jenkins Readme
- Added a Spinnaker Readme